RONDB-395: New retriable error codes
When a node failure happens, e.g. when running an OLTP RW transaction
in Sysbench, we will often report error code 4006. This error code
indicates overload and isn't necessarily leading to a retry by the
application logic. There are two main error categories on the MySQL
layer that leads to a retry. These are lock wait timeout and deadlock
detected. Thus all failures of transaction due to node failures should
be mapped to lock wait timeout.

The specific issue with 4006 is when we execute a transaction and we
need to perform a scan as part of this transaction. In this case we
need the scan to get a connection object from the same data node
where the transaction coordinator resides. If not possible the
transaction must fail. We will report a new error code 4042
that is mapped to lock wait timeout in this case.

We also map a node failure error from DBSPJ to lock wait timeout as
well.