RONDB-459: Fix for deactivate nodes
The method sendStopMgmd is used for both shutdown and stop of nodes.
When deactivating a node we stop the node after deactivating it.
In the case of deactivation of a MGM server we thus need to ensure
that we stop a node even if it is deactivated since it is part of
the deactivation. This required a new parameter to sendStopMgmd to
decide whether to handle stop of deactivated nodes or not.

Added checks that we don't stop deactivated nodes, they should already
be stopped.