RONDB-552: Make OS overhead configurable
Two new configuration parameters were introduced.
OsStaticOverhead and OsCpuOverhead

These variables are used to calculate OS overhead when calculating
automatic memory. When using AutomaticMemoryConfig there are two
modes, one is to set TotalMemoryConfig, in this case the user
has decided how much the total memory should be and this is used
and those parameters are ignored.

If the user has set to TotalMemoryConfig to 0 or simply not set it,
then the calculation will be based on the amount of memory in the
computer/VM/container it runs in. The default settings in this case
is intended for a setup in a VM where the data nodes can use the
entire VM except for some small processes used to monitor and change
the cluster.

However in some cases the user might need to run also some heavier
processes in their VM, in this case it is possible to increase
either OsStaticOverhead or OsCpuOverhead or both.

OsCpuOverhead is multiplied by the number of CPUs.

It is necessary to leave memory both for the OS itself and for
OS kernel buffers and for any applications that need to coexist
with RonDB.

The default setting on OsStaticOverhead is 1400M and the default
setting of OsCpuOverhead is 100M. In addition we will always avoid
using 1% of the memory size.

Thus by default a node with 16 VCPUs and 128GB of memory will
have OS overhead computed 1280M + 1400M + 16 * 100M = 4280M.
Thus we will almost 124GB of memory for the data node in this
case.

OsStaticOverhead cannot be decreased below 400M and OsCpuOverhead
cannot be decreased below 50M.