RONDB-555: Fix check for existing unique index in ndb_restore
RONDB-430 attempted to check whether any unique indexes exist in the
database that could cause trouble during restore and, if so, warn or
fail depending on whether --allow-unique-indexes is set. However,
instead of checking the database, it warned/failed on every index
present in the backup.

This patch fixes the check to actually check the database.